@@ -35,19 +35,20 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  flutter_custom_clippers: ^1.1.2
-  firebase_auth: ^0.18.4+1
-  firebase_core: ^0.5.3
-  cloud_firestore: ^0.14.4
-  progress_dialog: ^1.2.0
-  shared_preferences: ^0.5.12+4
-  google_maps_flutter: ^1.0.6
-  geolocator: ^6.1.13
-  location: ^3.1.0
-  geoflutterfire: ^2.2.1
   
-
+  cupertino_icons: ^1.0.2
+  flutter_custom_clippers: ^2.0.0
+  firebase_auth: ^1.0.0
+  firebase_core: ^1.0.0
+  cloud_firestore: ^1.0.0
+  progress_dialog: ^1.2.0
+  shared_preferences: ^2.0.0
+  google_maps_flutter: ^2.0.0
+  geolocator: ^7.0.0
+  location: ^4.0.0
+  geoflutterfire2: ^2.2.2
+  
+  
 dev_dependencies:
   flutter_test:
     sdk: flutter
