@@ -5,16 +5,16 @@ Client clientFromJson(String str) => Client.fromJson(json.decode(str));
 String clientToJson(Client data) => json.encode(data.toJson());
 
 class Client {
-  String id;
-  String username;
-  String email;
-  String password;
+  String? id;
+  String? username;
+  String? email;
+  String? password;
 
   Client({
-    required this.id,
-    required this.username,
-    required this.email,
-    required this.password,
+    this.id,
+    this.username,
+    this.email,
+    this.password,
   });
 
   factory Client.fromJson(Map<String, dynamic> json) => Client(
