@@ -1,72 +1,71 @@
 import 'dart:async';
+
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:geolocator/geolocator.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:location/location.dart' as location;
 import 'package:progress_dialog/progress_dialog.dart';
-import 'package:uber_clone_flutter_udemy/src/providers/auth_provider.dart';
-import 'package:uber_clone_flutter_udemy/src/providers/geofire_provider.dart';
-import 'package:uber_clone_flutter_udemy/src/providers/driver_provider.dart';
-import 'package:uber_clone_flutter_udemy/src/utils/my_progress_dialog.dart';
-import 'package:uber_clone_flutter_udemy/src/utils/snackbar.dart' as utils;
-import 'package:uber_clone_flutter_udemy/src/models/driver.dart';
+import 'package:uber_clone/src/utils/snackbar.dart' as utils;
 
+import '../../../models/driver.dart';
+import '../../../providers/auth_provider.dart';
+import '../../../providers/driver_provider.dart';
+import '../../../providers/geofire_provider.dart';
+import '../../../utils/my_progress_dialog.dart';
 
 class DriverMapController {
+  BuildContext? context;
+  Function? refresh;
+  GlobalKey<ScaffoldState> key = GlobalKey<ScaffoldState>();
+  final Completer<GoogleMapController> _mapController = Completer();
 
-  BuildContext context;
-  Function refresh;
-  GlobalKey<ScaffoldState> key = new GlobalKey<ScaffoldState>();
-  Completer<GoogleMapController> _mapController = Completer();
-
-  CameraPosition initialPosition = CameraPosition(
-    target: LatLng(1.2342774, -77.2645446),
-    zoom: 14.0
-  );
+  CameraPosition initialPosition =
+      const CameraPosition(target: LatLng(1.2342774, -77.2645446), zoom: 14.0);
 
   Map<MarkerId, Marker> markers = <MarkerId, Marker>{};
 
-  Position _position;
-  StreamSubscription<Position> _positionStream;
+  Position? _position;
+  StreamSubscription<Position>? _positionStream;
 
-  BitmapDescriptor markerDriver;
+  BitmapDescriptor? markerDriver;
 
-  GeofireProvider _geofireProvider;
-  AuthProvider _authProvider;
-  DriverProvider _driverProvider;
+  GeofireProvider? _geofireProvider;
+  AuthProvider? _authProvider;
+  DriverProvider? _driverProvider;
 
   bool isConnect = false;
-  ProgressDialog _progressDialog;
+  ProgressDialog? _progressDialog;
 
-  StreamSubscription<DocumentSnapshot> _statusSuscription;
-  StreamSubscription<DocumentSnapshot> _driverInfoSuscription;
+  StreamSubscription<DocumentSnapshot>? _statusSuscription;
+  StreamSubscription<DocumentSnapshot>? _driverInfoSuscription;
 
-  Driver driver;
+  Driver driver = Driver();
 
   Future init(BuildContext context, Function refresh) async {
     this.context = context;
     this.refresh = refresh;
-    _geofireProvider = new GeofireProvider();
-    _authProvider = new AuthProvider();
-    _driverProvider = new DriverProvider();
-    _progressDialog = MyProgressDialog.createProgressDialog(context, 'Conectandose...');
+    _geofireProvider = GeofireProvider();
+    _authProvider = AuthProvider();
+    _driverProvider = DriverProvider();
+    _progressDialog =
+        MyProgressDialog.createProgressDialog(context, 'Conectandose...');
     markerDriver = await createMarkerImageFromAsset('assets/img/taxi_icon.png');
     checkGPS();
     getDriverInfo();
   }
 
   void getDriverInfo() {
-    Stream<DocumentSnapshot> driverStream = _driverProvider.getByIdStream(_authProvider.getUser().uid);
+    Stream<DocumentSnapshot> driverStream =
+        _driverProvider!.getByIdStream(_authProvider!.getUser().uid);
     _driverInfoSuscription = driverStream.listen((DocumentSnapshot document) {
-      driver = Driver.fromJson(document.data());
-      refresh();
+      driver = Driver.fromJson(document.data() as Map<String, dynamic>);
+      refresh!();
     });
   }
 
   void openDrawer() {
-    key.currentState.openDrawer();
+    key.currentState!.openDrawer();
   }
 
   void dispose() {
@@ -76,102 +75,83 @@ class DriverMapController {
   }
 
   void signOut() async {
-    await _authProvider.signOut();
-    Navigator.pushNamedAndRemoveUntil(context, 'home', (route) => false);
+    await _authProvider!.signOut();
+    Navigator.pushNamedAndRemoveUntil(context!, 'home', (route) => false);
   }
 
   void onMapCreated(GoogleMapController controller) {
-    controller.setMapStyle('[{"elementType":"geometry","stylers":[{"color":"#212121"}]},{"elementType":"labels.icon","stylers":[{"visibility":"off"}]},{"elementType":"labels.text.fill","stylers":[{"color":"#757575"}]},{"elementType":"labels.text.stroke","stylers":[{"color":"#212121"}]},{"featureType":"administrative","elementType":"geometry","stylers":[{"color":"#757575"}]},{"featureType":"administrative.country","elementType":"labels.text.fill","stylers":[{"color":"#9e9e9e"}]},{"featureType":"administrative.land_parcel","stylers":[{"visibility":"off"}]},{"featureType":"administrative.locality","elementType":"labels.text.fill","stylers":[{"color":"#bdbdbd"}]},{"featureType":"poi","elementType":"labels.text.fill","stylers":[{"color":"#757575"}]},{"featureType":"poi.park","elementType":"geometry","stylers":[{"color":"#181818"}]},{"featureType":"poi.park","elementType":"labels.text.fill","stylers":[{"color":"#616161"}]},{"featureType":"poi.park","elementType":"labels.text.stroke","stylers":[{"color":"#1b1b1b"}]},{"featureType":"road","elementType":"geometry.fill","stylers":[{"color":"#2c2c2c"}]},{"featureType":"road","elementType":"labels.text.fill","stylers":[{"color":"#8a8a8a"}]},{"featureType":"road.arterial","elementType":"geometry","stylers":[{"color":"#373737"}]},{"featureType":"road.highway","elementType":"geometry","stylers":[{"color":"#3c3c3c"}]},{"featureType":"road.highway.controlled_access","elementType":"geometry","stylers":[{"color":"#4e4e4e"}]},{"featureType":"road.local","elementType":"labels.text.fill","stylers":[{"color":"#616161"}]},{"featureType":"transit","elementType":"labels.text.fill","stylers":[{"color":"#757575"}]},{"featureType":"water","elementType":"geometry","stylers":[{"color":"#000000"}]},{"featureType":"water","elementType":"labels.text.fill","stylers":[{"color":"#3d3d3d"}]}]');
+    controller.setMapStyle(
+        '[{"elementType":"geometry","stylers":[{"color":"#212121"}]},{"elementType":"labels.icon","stylers":[{"visibility":"off"}]},{"elementType":"labels.text.fill","stylers":[{"color":"#757575"}]},{"elementType":"labels.text.stroke","stylers":[{"color":"#212121"}]},{"featureType":"administrative","elementType":"geometry","stylers":[{"color":"#757575"}]},{"featureType":"administrative.country","elementType":"labels.text.fill","stylers":[{"color":"#9e9e9e"}]},{"featureType":"administrative.land_parcel","stylers":[{"visibility":"off"}]},{"featureType":"administrative.locality","elementType":"labels.text.fill","stylers":[{"color":"#bdbdbd"}]},{"featureType":"poi","elementType":"labels.text.fill","stylers":[{"color":"#757575"}]},{"featureType":"poi.park","elementType":"geometry","stylers":[{"color":"#181818"}]},{"featureType":"poi.park","elementType":"labels.text.fill","stylers":[{"color":"#616161"}]},{"featureType":"poi.park","elementType":"labels.text.stroke","stylers":[{"color":"#1b1b1b"}]},{"featureType":"road","elementType":"geometry.fill","stylers":[{"color":"#2c2c2c"}]},{"featureType":"road","elementType":"labels.text.fill","stylers":[{"color":"#8a8a8a"}]},{"featureType":"road.arterial","elementType":"geometry","stylers":[{"color":"#373737"}]},{"featureType":"road.highway","elementType":"geometry","stylers":[{"color":"#3c3c3c"}]},{"featureType":"road.highway.controlled_access","elementType":"geometry","stylers":[{"color":"#4e4e4e"}]},{"featureType":"road.local","elementType":"labels.text.fill","stylers":[{"color":"#616161"}]},{"featureType":"transit","elementType":"labels.text.fill","stylers":[{"color":"#757575"}]},{"featureType":"water","elementType":"geometry","stylers":[{"color":"#000000"}]},{"featureType":"water","elementType":"labels.text.fill","stylers":[{"color":"#3d3d3d"}]}]');
     _mapController.complete(controller);
   }
 
   void saveLocation() async {
-    await _geofireProvider.create(
-        _authProvider.getUser().uid,
-        _position.latitude,
-        _position.longitude
-    );
-    _progressDialog.hide();
+    await _geofireProvider!.create(_authProvider!.getUser().uid,
+        _position!.latitude, _position!.longitude);
+    _progressDialog!.hide();
   }
 
   void connect() {
     if (isConnect) {
       disconnect();
-    }
-    else {
-      _progressDialog.show();
+    } else {
+      _progressDialog!.show();
       updateLocation();
     }
   }
 
   void disconnect() {
     _positionStream?.cancel();
-    _geofireProvider.delete(_authProvider.getUser().uid);
+    _geofireProvider!.delete(_authProvider!.getUser().uid);
   }
 
   void checkIfIsConnect() {
     Stream<DocumentSnapshot> status =
-    _geofireProvider.getLocationByIdStream(_authProvider.getUser().uid);
+        _geofireProvider!.getLocationByIdStream(_authProvider!.getUser().uid);
 
     _statusSuscription = status.listen((DocumentSnapshot document) {
       if (document.exists) {
         isConnect = true;
-      }
-      else {
+      } else {
         isConnect = false;
       }
 
-      refresh();
+      refresh!();
     });
-
   }
 
-  void updateLocation() async  {
+  void updateLocation() async {
     try {
       await _determinePosition();
       _position = await Geolocator.getLastKnownPosition();
       centerPosition();
       saveLocation();
 
-      addMarker(
-          'driver',
-          _position.latitude,
-          _position.longitude,
-          'Tu posicion',
-          '',
-          markerDriver
-      );
-      refresh();
+      addMarker('driver', _position!.latitude, _position!.longitude,
+          'Tu posicion', '', markerDriver!);
+      refresh!();
 
       _positionStream = Geolocator.getPositionStream(
-          desiredAccuracy: LocationAccuracy.best,
-          distanceFilter: 1
-      ).listen((Position position) {
-          _position = position;
-          addMarker(
-              'driver',
-              _position.latitude,
-              _position.longitude,
-              'Tu posicion',
-              '',
-              markerDriver
-          );
-          animateCameraToPosition(_position.latitude, _position.longitude);
-          saveLocation();
-          refresh();
+              desiredAccuracy: LocationAccuracy.best, distanceFilter: 1)
+          .listen((Position position) {
+        _position = position;
+        addMarker('driver', _position!.latitude, _position!.longitude,
+            'Tu posicion', '', markerDriver!);
+        animateCameraToPosition(_position!.latitude, _position!.longitude);
+        saveLocation();
+        refresh!();
       });
-
-    } catch(error) {
+    } catch (error) {
       print('Error en la localizacion: $error');
     }
   }
 
   void centerPosition() {
     if (_position != null) {
-      animateCameraToPosition(_position.latitude, _position.longitude);
-    }
-    else {
-      utils.Snackbar.showSnackbar(context, key, 'Activa el GPS para obtener la posicion');
+      animateCameraToPosition(_position!.latitude, _position!.longitude);
+    } else {
+      utils.Snackbar.showSnackbar(
+          context!, key, 'Activa el GPS para obtener la posicion');
     }
   }
 
@@ -181,8 +161,7 @@ class DriverMapController {
       print('GPS ACTIVADO');
       updateLocation();
       checkIfIsConnect();
-    }
-    else {
+    } else {
       print('GPS DESACTIVADO');
       bool locationGPS = await location.Location().requestService();
       if (locationGPS) {
@@ -191,7 +170,6 @@ class DriverMapController {
         print('ACTIVO EL GPS');
       }
     }
-
   }
 
   Future<Position> _determinePosition() async {
@@ -223,49 +201,31 @@ class DriverMapController {
 
   Future animateCameraToPosition(double latitude, double longitude) async {
     GoogleMapController controller = await _mapController.future;
-    if (controller != null) {
-      controller.animateCamera(CameraUpdate.newCameraPosition(
-        CameraPosition(
-          bearing: 0,
-          target: LatLng(latitude, longitude),
-          zoom: 17
-        )
-      ));
-    }
+    controller.animateCamera(CameraUpdate.newCameraPosition(CameraPosition(
+        bearing: 0, target: LatLng(latitude, longitude), zoom: 17)));
   }
 
   Future<BitmapDescriptor> createMarkerImageFromAsset(String path) async {
-    ImageConfiguration configuration = ImageConfiguration();
+    ImageConfiguration configuration = const ImageConfiguration();
     BitmapDescriptor bitmapDescriptor =
         await BitmapDescriptor.fromAssetImage(configuration, path);
     return bitmapDescriptor;
   }
 
-  void addMarker(
-      String markerId,
-      double lat,
-      double lng,
-      String title,
-      String content,
-      BitmapDescriptor iconMarker
-    ) {
-
+  void addMarker(String markerId, double lat, double lng, String title,
+      String content, BitmapDescriptor iconMarker) {
     MarkerId id = MarkerId(markerId);
     Marker marker = Marker(
-      markerId: id,
-      icon: iconMarker,
-      position: LatLng(lat, lng),
-      infoWindow: InfoWindow(title: title, snippet: content),
-      draggable: false,
-      zIndex: 2,
-      flat: true,
-      anchor: Offset(0.5, 0.5),
-      rotation: _position.heading
-    );
+        markerId: id,
+        icon: iconMarker,
+        position: LatLng(lat, lng),
+        infoWindow: InfoWindow(title: title, snippet: content),
+        draggable: false,
+        zIndex: 2,
+        flat: true,
+        anchor: const Offset(0.5, 0.5),
+        rotation: _position!.heading);
 
     markers[id] = marker;
-
   }
-
-
 }
