@@ -2,19 +2,18 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 
 class AuthProvider {
-
-  FirebaseAuth _firebaseAuth;
+  FirebaseAuth? _firebaseAuth;
 
   AuthProvider() {
     _firebaseAuth = FirebaseAuth.instance;
   }
 
   User getUser() {
-    return _firebaseAuth.currentUser;
+    return _firebaseAuth!.currentUser!;
   }
 
   bool isSignedIn() {
-    final currentUser = _firebaseAuth.currentUser;
+    final currentUser = _firebaseAuth!.currentUser;
 
     if (currentUser == null) {
       return false;
@@ -26,18 +25,16 @@ class AuthProvider {
   void checkIfUserIsLogged(BuildContext context, String typeUser) {
     FirebaseAuth.instance.authStateChanges().listen((User user) {
       // QUE EL USUARIO ESTA LOGEADO
-      if (user != null && typeUser != null) {
-
+      if (typeUser != null) {
         if (typeUser == 'client') {
-          Navigator.pushNamedAndRemoveUntil(context, 'client/map', (route) => false);
-        }
-        else {
-          Navigator.pushNamedAndRemoveUntil(context, 'driver/map', (route) => false);
+          Navigator.pushNamedAndRemoveUntil(
+              context, 'client/map', (route) => false);
+        } else {
+          Navigator.pushNamedAndRemoveUntil(
+              context, 'driver/map', (route) => false);
         }
         print('El usuario esta logeado');
-
-      }
-      else {
+      } else {
         print('El usuario no esta logeado');
       }
     });
@@ -47,48 +44,37 @@ class AuthProvider {
     String errorMessage;
 
     try {
-
-      await _firebaseAuth.signInWithEmailAndPassword(email: email, password: password);
-
-    } catch(error) {
+      await _firebaseAuth!
+          .signInWithEmailAndPassword(email: email, password: password);
+    } catch (error) {
       print(error);
       // CORREO INVALIDO
       // PASSWORD INCORRECTO
       // NO HAY CONEXION A INTERNET
-      errorMessage = error.code;
+      errorMessage = error.toString();
     }
 
-    if (errorMessage != null) {
-      return Future.error(errorMessage);
-    }
-
-    return true;
+    return Future.error(errorMessage);
   }
 
   Future<bool> register(String email, String password) async {
     String errorMessage;
 
     try {
-
-      await _firebaseAuth.createUserWithEmailAndPassword(email: email, password: password);
-
-    } catch(error) {
+      await _firebaseAuth!
+          .createUserWithEmailAndPassword(email: email, password: password);
+    } catch (error) {
       print(error);
       // CORREO INVALIDO
       // PASSWORD INCORRECTO
       // NO HAY CONEXION A INTERNET
-      errorMessage = error.code;
+      errorMessage = error.toString();
     }
 
-    if (errorMessage != null) {
-      return Future.error(errorMessage);
-    }
-
-    return true;
+    return Future.error(errorMessage);
   }
 
-  Future<void> signOut() async {
-    return Future.wait([_firebaseAuth.signOut()]);
+  Future<Future<List<void>>> signOut() async {
+    return Future.wait([_firebaseAuth!.signOut()]);
   }
-
 }
