@@ -1,43 +1,41 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:uber_clone_flutter_udemy/src/models/client.dart';
+
+import '../models/client.dart';
 
 class ClientProvider {
-
-  CollectionReference _ref;
+  CollectionReference? _ref;
 
   ClientProvider() {
     _ref = FirebaseFirestore.instance.collection('Clients');
   }
 
-  Future<void> create(Client client) {
+  Future<void>? create(Client client) {
     String errorMessage;
 
     try {
-      return _ref.doc(client.id).set(client.toJson());
-    } catch(error) {
-      errorMessage = error.code;
+      return _ref!.doc(client.id).set(client.toJson());
+    } catch (error) {
+      errorMessage = error.toString();
     }
 
     if (errorMessage != null) {
       return Future.error(errorMessage);
     }
+    return null;
   }
 
   Stream<DocumentSnapshot> getByIdStream(String id) {
-    return _ref.doc(id).snapshots(includeMetadataChanges: true);
+    return _ref!.doc(id).snapshots(includeMetadataChanges: true);
   }
 
-  Future<Client> getById(String id) async {
-    DocumentSnapshot document = await _ref.doc(id).get();
+  Future<Client?> getById(String id) async {
+    DocumentSnapshot document = await _ref!.doc(id).get();
 
     if (document.exists) {
-      Client client = Client.fromJson(document.data());
+      Client client = Client.fromJson(document.data() as Map<String, dynamic>);
       return client;
     }
 
     return null;
-
   }
-
-
 }
