@@ -1,21 +1,22 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'package:uber_clone_flutter_udemy/src/pages/driver/map/driver_map_controller.dart';
-import 'package:uber_clone_flutter_udemy/src/widgets/button_app.dart';
+
+import '../../../widgets/button_app.dart';
+import 'driver_map_controller.dart';
 
 class DriverMapPage extends StatefulWidget {
+  const DriverMapPage({super.key});
+
   @override
   _DriverMapPageState createState() => _DriverMapPageState();
 }
 
 class _DriverMapPageState extends State<DriverMapPage> {
-
-  DriverMapController _con = new DriverMapController();
+  final DriverMapController _con = DriverMapController();
 
   @override
   void initState() {
-    // TODO: implement initState
     super.initState();
 
     SchedulerBinding.instance.addPostFrameCallback((timeStamp) {
@@ -33,7 +34,6 @@ class _DriverMapPageState extends State<DriverMapPage> {
 
   @override
   Widget build(BuildContext context) {
-
     return Scaffold(
       key: _con.key,
       drawer: _drawer(),
@@ -66,6 +66,7 @@ class _DriverMapPageState extends State<DriverMapPage> {
         padding: EdgeInsets.zero,
         children: [
           DrawerHeader(
+            decoration: const BoxDecoration(color: Colors.amber),
             child: Column(
               mainAxisAlignment: MainAxisAlignment.start,
               crossAxisAlignment: CrossAxisAlignment.start,
@@ -73,45 +74,40 @@ class _DriverMapPageState extends State<DriverMapPage> {
                 Container(
                   child: Text(
                     _con.driver?.username ?? 'Nombre de usuario',
-                    style: TextStyle(
-                      fontSize: 18,
-                      color: Colors.black,
-                      fontWeight: FontWeight.bold
-                    ),
+                    style: const TextStyle(
+                        fontSize: 18,
+                        color: Colors.black,
+                        fontWeight: FontWeight.bold),
                     maxLines: 1,
                   ),
                 ),
                 Container(
                   child: Text(
-                    _con.driver?.email ?? 'Correo electronico' ,
+                    _con.driver?.email ?? 'Correo electronico',
                     style: TextStyle(
                         fontSize: 14,
                         color: Colors.grey[800],
-                        fontWeight: FontWeight.bold
-                    ),
+                        fontWeight: FontWeight.bold),
                     maxLines: 1,
                   ),
                 ),
-                SizedBox(height: 10),
-                CircleAvatar(
+                const SizedBox(height: 10),
+                const CircleAvatar(
                   backgroundImage: AssetImage('assets/img/profile.jpg'),
                   radius: 40,
                 )
               ],
             ),
-            decoration: BoxDecoration(
-              color: Colors.amber
-            ),
           ),
           ListTile(
-            title: Text('Editar perfil'),
-            trailing: Icon(Icons.edit),
+            title: const Text('Editar perfil'),
+            trailing: const Icon(Icons.edit),
             // leading: Icon(Icons.cancel),
             onTap: () {},
           ),
           ListTile(
-            title: Text('Cerrar sesion'),
-            trailing: Icon(Icons.power_settings_new),
+            title: const Text('Cerrar sesion'),
+            trailing: const Icon(Icons.power_settings_new),
             // leading: Icon(Icons.cancel),
             onTap: _con.signOut,
           ),
@@ -125,13 +121,13 @@ class _DriverMapPageState extends State<DriverMapPage> {
       onTap: _con.centerPosition,
       child: Container(
         alignment: Alignment.centerRight,
-        margin: EdgeInsets.symmetric(horizontal: 5),
+        margin: const EdgeInsets.symmetric(horizontal: 5),
         child: Card(
-          shape: CircleBorder(),
+          shape: const CircleBorder(),
           color: Colors.white,
           elevation: 4.0,
           child: Container(
-            padding: EdgeInsets.all(10),
+            padding: const EdgeInsets.all(10),
             child: Icon(
               Icons.location_searching,
               color: Colors.grey[600],
@@ -145,23 +141,26 @@ class _DriverMapPageState extends State<DriverMapPage> {
 
   Widget _buttonDrawer() {
     return Container(
-        alignment: Alignment.centerLeft,
-        child: IconButton(
-          onPressed: _con.openDrawer,
-          icon: Icon(Icons.menu, color: Colors.white,),
+      alignment: Alignment.centerLeft,
+      child: IconButton(
+        onPressed: _con.openDrawer,
+        icon: const Icon(
+          Icons.menu,
+          color: Colors.white,
         ),
-      );
+      ),
+    );
   }
 
   Widget _buttonConnect() {
     return Container(
       height: 50,
       alignment: Alignment.bottomCenter,
-      margin: EdgeInsets.symmetric(horizontal: 60, vertical: 30),
+      margin: const EdgeInsets.symmetric(horizontal: 60, vertical: 30),
       child: ButtonApp(
         onPressed: _con.connect,
         text: _con.isConnect ? 'DESCONECTARSE' : 'CONECTARSE',
-        color: _con.isConnect ? Colors.grey[300] : Colors.amber,
+        color: _con.isConnect ? Colors.grey[300]! : Colors.amber,
         textColor: Colors.black,
       ),
     );
@@ -181,5 +180,4 @@ class _DriverMapPageState extends State<DriverMapPage> {
   void refresh() {
     setState(() {});
   }
-
 }
