@@ -1,9 +1,9 @@
-import 'package:shared_preferences/shared_preferences.dart';
 import 'dart:convert';
 
-class SharedPref {
+import 'package:shared_preferences/shared_preferences.dart';
 
-  void save (String key, String value) async {
+class SharedPref {
+  void save(String key, String value) async {
     final prefs = await SharedPreferences.getInstance();
     prefs.setString(key, json.encode(value));
   }
@@ -13,7 +13,7 @@ class SharedPref {
 
     if (prefs.getString(key) == null) return null;
 
-    return json.decode(prefs.getString(key));
+    return json.decode(prefs.getString(key).toString());
   }
 
   // Nombre - true - false
@@ -27,5 +27,4 @@ class SharedPref {
     final prefs = await SharedPreferences.getInstance();
     return prefs.remove(key);
   }
-
 }
